requirements.txt
openai==1.3.0
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
PyPDF2==3.0.1
azure-storage-blob==12.19.0
gunicorn==21.2.0
